requirements.txt
Flask
openai-whisper
pyannote.audio
